reformat the code

+++ vit_pytorch/ats_vit.py
@@ -6,24 +6,29 @@ from torch import nn, einsum
 from einops import rearrange, repeat
 from einops.layers.torch import Rearrange
 
+
 # helpers
 
 def exists(val):
     return val is not None
 
+
 def pair(t):
     return t if isinstance(t, tuple) else (t, t)
 
+
 # adaptive token sampling functions and classes
 
-def log(t, eps = 1e-6):
+def log(t, eps=1e-6):
     return torch.log(t + eps)
 
-def sample_gumbel(shape, device, dtype, eps = 1e-6):
-    u = torch.empty(shape, device = device, dtype = dtype).uniform_(0, 1)
+
+def sample_gumbel(shape, device, dtype, eps=1e-6):
+    u = torch.empty(shape, device=device, dtype=dtype).uniform_(0, 1)
     return -log(-log(u, eps), eps)
 
-def batched_index_select(values, indices, dim = 1):
+
+def batched_index_select(values, indices, dim=1):
     value_dims = values.shape[(dim + 1):]
     values_shape, indices_shape = map(lambda t: list(t.shape), (values, indices))
     indices = indices[(..., *((None,) * len(value_dims)))]
@@ -39,8 +44,9 @@ def batched_index_select(values, indices, dim = 1):
     dim += value_expand_len
     return values.gather(dim, indices)
 
+
 class AdaptiveTokenSampling(nn.Module):
-    def __init__(self, output_num_tokens, eps = 1e-6):
+    def __init__(self, output_num_tokens, eps=1e-6):
         super().__init__()
         self.eps = eps
         self.output_num_tokens = output_num_tokens
@@ -54,7 +60,7 @@ class AdaptiveTokenSampling(nn.Module):
 
         # calculate the norms of the values, for weighting the scores, as described in the paper
 
-        value_norms = value[..., 1:, :].norm(dim = -1)
+        value_norms = value[..., 1:, :].norm(dim=-1)
 
         # weigh the attention scores by the norm of the values, sum across all heads
 
@@ -62,7 +68,7 @@ class AdaptiveTokenSampling(nn.Module):
 
         # normalize to 1
 
-        normed_cls_attn = cls_attn / (cls_attn.sum(dim = -1, keepdim = True) + eps)
+        normed_cls_attn = cls_attn / (cls_attn.sum(dim=-1, keepdim=True) + eps)
 
         # instead of using inverse transform sampling, going to invert the softmax and use gumbel-max sampling instead
 
@@ -76,17 +82,17 @@ class AdaptiveTokenSampling(nn.Module):
 
         # expand k times, k being the adaptive sampling number
 
-        pseudo_logits = repeat(pseudo_logits, 'b n -> b k n', k = output_num_tokens)
-        pseudo_logits = pseudo_logits + sample_gumbel(pseudo_logits.shape, device = device, dtype = dtype)
+        pseudo_logits = repeat(pseudo_logits, 'b n -> b k n', k=output_num_tokens)
+        pseudo_logits = pseudo_logits + sample_gumbel(pseudo_logits.shape, device=device, dtype=dtype)
 
         # gumble-max and add one to reserve 0 for padding / mask
 
-        sampled_token_ids = pseudo_logits.argmax(dim = -1) + 1
+        sampled_token_ids = pseudo_logits.argmax(dim=-1) + 1
 
         # calculate unique using torch.unique and then pad the sequence from the right
 
-        unique_sampled_token_ids_list = [torch.unique(t, sorted = True) for t in torch.unbind(sampled_token_ids)]
-        unique_sampled_token_ids = pad_sequence(unique_sampled_token_ids_list, batch_first = True)
+        unique_sampled_token_ids_list = [torch.unique(t, sorted=True) for t in torch.unbind(sampled_token_ids)]
+        unique_sampled_token_ids = pad_sequence(unique_sampled_token_ids_list, batch_first=True)
 
         # calculate the new mask, based on the padding
 
@@ -94,19 +100,20 @@ class AdaptiveTokenSampling(nn.Module):
 
         # CLS token never gets masked out (gets a value of True)
 
-        new_mask = F.pad(new_mask, (1, 0), value = True)
+        new_mask = F.pad(new_mask, (1, 0), value=True)
 
         # prepend a 0 token id to keep the CLS attention scores
 
-        unique_sampled_token_ids = F.pad(unique_sampled_token_ids, (1, 0), value = 0)
-        expanded_unique_sampled_token_ids = repeat(unique_sampled_token_ids, 'b n -> b h n', h = heads)
+        unique_sampled_token_ids = F.pad(unique_sampled_token_ids, (1, 0), value=0)
+        expanded_unique_sampled_token_ids = repeat(unique_sampled_token_ids, 'b n -> b h n', h=heads)
 
         # gather the new attention scores
 
-        new_attn = batched_index_select(attn, expanded_unique_sampled_token_ids, dim = 2)
+        new_attn = batched_index_select(attn, expanded_unique_sampled_token_ids, dim=2)
 
         # return the sampled attention scores, new mask (denoting padding), as well as the sampled token indices (for the residual)
         return new_attn, new_mask, unique_sampled_token_ids
+
 
 # classes
 
@@ -115,11 +122,13 @@ class PreNorm(nn.Module):
         super().__init__()
         self.norm = nn.LayerNorm(dim)
         self.fn = fn
+
     def forward(self, x, **kwargs):
         return self.fn(self.norm(x), **kwargs)
 
+
 class FeedForward(nn.Module):
-    def __init__(self, dim, hidden_dim, dropout = 0.):
+    def __init__(self, dim, hidden_dim, dropout=0.):
         super().__init__()
         self.net = nn.Sequential(
             nn.Linear(dim, hidden_dim),
@@ -128,20 +137,22 @@ class FeedForward(nn.Module):
             nn.Linear(hidden_dim, dim),
             nn.Dropout(dropout)
         )
+
     def forward(self, x):
         return self.net(x)
 
+
 class Attention(nn.Module):
-    def __init__(self, dim, heads = 8, dim_head = 64, dropout = 0., output_num_tokens = None):
+    def __init__(self, dim, heads=8, dim_head=64, dropout=0., output_num_tokens=None):
         super().__init__()
-        inner_dim = dim_head *  heads
+        inner_dim = dim_head * heads
         self.heads = heads
         self.scale = dim_head ** -0.5
 
-        self.attend = nn.Softmax(dim = -1)
+        self.attend = nn.Softmax(dim=-1)
         self.dropout = nn.Dropout(dropout)
 
-        self.to_qkv = nn.Linear(dim, inner_dim * 3, bias = False)
+        self.to_qkv = nn.Linear(dim, inner_dim * 3, bias=False)
 
         self.output_num_tokens = output_num_tokens
         self.ats = AdaptiveTokenSampling(output_num_tokens) if exists(output_num_tokens) else None
@@ -154,8 +165,8 @@ class Attention(nn.Module):
     def forward(self, x, *, mask):
         num_tokens = x.shape[1]
 
-        qkv = self.to_qkv(x).chunk(3, dim = -1)
-        q, k, v = map(lambda t: rearrange(t, 'b n (h d) -> b h n d', h = self.heads), qkv)
+        qkv = self.to_qkv(x).chunk(3, dim=-1)
+        q, k, v = map(lambda t: rearrange(t, 'b n (h d) -> b h n d', h=self.heads), qkv)
 
         dots = torch.matmul(q, k.transpose(-1, -2)) * self.scale
 
@@ -172,25 +183,26 @@ class Attention(nn.Module):
         # if adaptive token sampling is enabled
         # and number of tokens is greater than the number of output tokens
         if exists(self.output_num_tokens) and (num_tokens - 1) > self.output_num_tokens:
-            attn, mask, sampled_token_ids = self.ats(attn, v, mask = mask)
+            attn, mask, sampled_token_ids = self.ats(attn, v, mask=mask)
 
         out = torch.matmul(attn, v)
         out = rearrange(out, 'b h n d -> b n (h d)')
 
         return self.to_out(out), mask, sampled_token_ids
 
+
 class Transformer(nn.Module):
-    def __init__(self, dim, depth, max_tokens_per_depth, heads, dim_head, mlp_dim, dropout = 0.):
+    def __init__(self, dim, depth, max_tokens_per_depth, heads, dim_head, mlp_dim, dropout=0.):
         super().__init__()
         assert len(max_tokens_per_depth) == depth, 'max_tokens_per_depth must be a tuple of length that is equal to the depth of the transformer'
-        assert sorted(max_tokens_per_depth, reverse = True) == list(max_tokens_per_depth), 'max_tokens_per_depth must be in decreasing order'
+        assert sorted(max_tokens_per_depth, reverse=True) == list(max_tokens_per_depth), 'max_tokens_per_depth must be in decreasing order'
         assert min(max_tokens_per_depth) > 0, 'max_tokens_per_depth must have at least 1 token at any layer'
 
         self.layers = nn.ModuleList([])
         for _, output_num_tokens in zip(range(depth), max_tokens_per_depth):
             self.layers.append(nn.ModuleList([
-                PreNorm(dim, Attention(dim, output_num_tokens = output_num_tokens, heads = heads, dim_head = dim_head, dropout = dropout)),
-                PreNorm(dim, FeedForward(dim, mlp_dim, dropout = dropout))
+                PreNorm(dim, Attention(dim, output_num_tokens=output_num_tokens, heads=heads, dim_head=dim_head, dropout=dropout)),
+                PreNorm(dim, FeedForward(dim, mlp_dim, dropout=dropout))
             ]))
 
     def forward(self, x):
@@ -198,18 +210,18 @@ class Transformer(nn.Module):
 
         # use mask to keep track of the paddings when sampling tokens
         # as the duplicates (when sampling) are just removed, as mentioned in the paper
-        mask = torch.ones((b, n), device = device, dtype = torch.bool)
+        mask = torch.ones((b, n), device=device, dtype=torch.bool)
 
-        token_ids = torch.arange(n, device = device)
-        token_ids = repeat(token_ids, 'n -> b n', b = b)
+        token_ids = torch.arange(n, device=device)
+        token_ids = repeat(token_ids, 'n -> b n', b=b)
 
         for attn, ff in self.layers:
-            attn_out, mask, sampled_token_ids = attn(x, mask = mask)
+            attn_out, mask, sampled_token_ids = attn(x, mask=mask)
 
             # when token sampling, one needs to then gather the residual tokens with the sampled token ids
             if exists(sampled_token_ids):
-                x = batched_index_select(x, sampled_token_ids, dim = 1)
-                token_ids = batched_index_select(token_ids, sampled_token_ids, dim = 1)
+                x = batched_index_select(x, sampled_token_ids, dim=1)
+                token_ids = batched_index_select(token_ids, sampled_token_ids, dim=1)
 
             x = x + attn_out
 
@@ -217,8 +229,9 @@ class Transformer(nn.Module):
 
         return x, token_ids
 
+
 class ViT(nn.Module):
-    def __init__(self, *, image_size, patch_size, num_classes, dim, depth, max_tokens_per_depth, heads, mlp_dim, channels = 3, dim_head = 64, dropout = 0., emb_dropout = 0.):
+    def __init__(self, *, image_size, patch_size, num_classes, dim, depth, max_tokens_per_depth, heads, mlp_dim, channels=3, dim_head=64, dropout=0., emb_dropout=0.):
         super().__init__()
         image_height, image_width = pair(image_size)
         patch_height, patch_width = pair(patch_size)
@@ -229,7 +242,7 @@ class ViT(nn.Module):
         patch_dim = channels * patch_height * patch_width
 
         self.to_patch_embedding = nn.Sequential(
-            Rearrange('b c (h p1) (w p2) -> b (h w) (p1 p2 c)', p1 = patch_height, p2 = patch_width),
+            Rearrange('b c (h p1) (w p2) -> b (h w) (p1 p2 c)', p1=patch_height, p2=patch_width),
             nn.Linear(patch_dim, dim),
         )
 
@@ -244,11 +257,11 @@ class ViT(nn.Module):
             nn.Linear(dim, num_classes)
         )
 
-    def forward(self, img, return_sampled_token_ids = False):
+    def forward(self, img, return_sampled_token_ids=False):
         x = self.to_patch_embedding(img)
         b, n, _ = x.shape
 
-        cls_tokens = repeat(self.cls_token, '() n d -> b n d', b = b)
+        cls_tokens = repeat(self.cls_token, '() n d -> b n d', b=b)
         x = torch.cat((cls_tokens, x), dim=1)
         x += self.pos_embedding[:, :(n + 1)]
         x = self.dropout(x)
